modified: draftcontroller gets cleaned
