+++ backend/src/main/resources/application.properties
@@ -7,7 +7,7 @@ spring.security.oauth2.client.registration.google.client-id=${GOOGLE_CLIENT_ID}
 spring.security.oauth2.client.registration.google.client-secret=${GOOGLE_CLIENT_SECRET}
 spring.security.oauth2.client.registration.google.scope=profile,email
 
-spring.datasource.url=jdbc:mysql://20.187.149.238:3306/${MYSQL_DATABASE}
+spring.datasource.url=jdbc:mysql://localhost/${MYSQL_DATABASE}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.username=${MYSQL_USERNAME}
 spring.datasource.password=${MYSQL_PASSWORD}
